Factorize out the buffer accessor view in its own class
